+++ CHANGELOG.rst
@@ -1,3 +1,11 @@
+DEV
+===
+
+* Cart modifiers can add additional text ``data`` beside ``label`` and
+  ``value`` (this data would be saved in ExtraOrderPriceField and
+  ExtraOrderItemPriceField models). This allows filtering extra order and
+  extra order item fields by some criteria.
+
 Version 0.0.13
 ==============